+++ pyogrio/geopandas.py
@@ -403,10 +403,10 @@ def write_dataframe(
     metadata : dict, optional (default: None)
         alias of layer_metadata
     dataset_options : dict, optional
-        Dataset creation option (format specific) passed to OGR. Specify as
+        Dataset creation options (format specific) passed to OGR. Specify as
         a key-value dictionary.
     layer_options : dict, optional
-        Layer creation option (format specific) passed to OGR. Specify as
+        Layer creation options (format specific) passed to OGR. Specify as
         a key-value dictionary.
     **kwargs
         Additional driver-specific dataset or layer creation options passed
@@ -496,9 +496,7 @@ def write_dataframe(
         # If there is data, infer layer geometry type + promote_to_multi
         if not df.empty:
             # None/Empty geometries sometimes report as Z incorrectly, so ignore them
-            has_z_arr = geometry[
-                (geometry != np.array(None)) & (~geometry.is_empty)
-            ].has_z
+            has_z_arr = geometry[geometry.notna() & (~geometry.is_empty)].has_z
             has_z = has_z_arr.any()
             all_z = has_z_arr.all()